requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
gunicorn==21.2.0
email-validator==2.0.0